requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
loguru
requests
streamlit
plotly
PyMuPDF
python-docx
scikit-learn
numpy
rapidfuzz
spacy
langchain
langchain-openai
langgraph
langsmith
reportlab